[composables] ARCH-22 (1D.9): first slice of the annotation.js split
annotation.js is 1700 lines and owns everything from canvas lifecycle to
persistence; start splitting it incrementally without touching its external
API. This pass extracts one module: the drawing-tools configuration.

useDrawingTools (src/composables/players/drawingTools.js) now owns the
pencil / eraser / shape / text tool state transitions and brush
configuration: onAnnotateClicked, onEraseClicked, _resetEraserWidth,
_resetColor, _resetPencil, resetPencilConfiguration, the pencil/text
preference handlers (onChangePencilColor / onChangePencilWidth /
onChangeTextColor), toggleShapeMode, setShapeTool and
setAnnotationDrawingMode. Bodies moved verbatim; the PSBrush, EraserBrush
and lockBrushToFirstPointer imports moved with them. The shared refs and
helpers the bodies close over travel through a plain ctx object
(fabricCanvas, isEraserModeOn, isShapeMode, currentShape, pencilColor,
pencilWidth, textColor, showCanvas, localPreferences).

annotation.js stays the facade: it calls useDrawingTools after showCanvas
is declared (still before the canvas-mirror watchers that run
configureCanvas → _resetPencil) and re-exposes the same names in its
return object, so consumers see no difference. The isShapeMode
selection-suppression watcher stays in annotation.js — it is canvas-state
wiring, not a tool transition. All targeted functions moved cleanly; none
had to be left behind.
