+++ src/composables/players/annotation.js
@@ -11,26 +11,22 @@ import {
   getFabricDocument,
   util
 } from 'fabric'
-import { PSStroke, PSBrush } from 'fabricjs-psbrush'
+import { PSStroke } from 'fabricjs-psbrush'
 import moment from 'moment'
 import { v4 as uuidv4 } from 'uuid'
 import { markRaw, ref, watch } from 'vue'
 
+import { useDrawingTools } from '@/composables/players/drawingTools'
 import {
   SHAPE_WIDTHS,
   addSerialization,
   attachShapeDrawing,
   buildReadOnlyShape,
   deserializePSStroke,
-  getAnnotationContainMapping,
-  lockBrushToFirstPointer
+  getAnnotationContainMapping
 } from '@/lib/players/annotation'
 import { normalizeType } from '@/lib/players/annotationTypes'
-import {
-  Eraser,
-  EraserBrush,
-  reviveObjectEraser
-} from '@/lib/players/eraserbrush'
+import { Eraser, reviveObjectEraser } from '@/lib/players/eraserbrush'
 import clipboard from '@/lib/clipboard'
 import { formatFullDate } from '@/lib/time'
 import localPreferences from '@/lib/preferences'
@@ -724,109 +720,6 @@ export const useAnnotation = ({
 
   // Events
 
-  const onChangePencilColor = color => {
-    pencilColor.value = color
-    _resetColor()
-    localPreferences.setPreference('player:pencil-color', pencilColor.value)
-  }
-
-  const onChangePencilWidth = pencil => {
-    pencilWidth.value = pencil
-    _resetPencil()
-    localPreferences.setPreference('player:pencil-width', pencilWidth.value)
-  }
-
-  const onChangeTextColor = newValue => {
-    textColor.value = newValue
-    localPreferences.setPreference('player:text-color', textColor.value)
-  }
-
-  const _resetColor = () => {
-    if (!fabricCanvas.value) return
-    fabricCanvas.value.freeDrawingBrush.color = pencilColor.value
-  }
-
-  const _resetPencil = () => {
-    if (!fabricCanvas.value) return
-    const converter = {
-      huge: 30,
-      big: 20,
-      medium: 10,
-      small: 4,
-      tiny: 2
-    }
-    const strokeWidth = converter[pencilWidth.value]
-    fabricCanvas.value.freeDrawingBrush.width = strokeWidth
-  }
-
-  const resetPencilConfiguration = () => {
-    pencilColor.value =
-      localPreferences.getPreference('player:pencil-color') || '#ff3860'
-    textColor.value =
-      localPreferences.getPreference('player:text-color') || '#ff3860'
-    pencilWidth.value =
-      localPreferences.getPreference('player:pencil-width') || 'big'
-
-    _resetColor()
-    _resetPencil()
-  }
-
-  // Drawing config
-
-  const onAnnotateClicked = () => {
-    showCanvas()
-    // Toggle off only when the pencil itself is the active brush. If the
-    // eraser is on, switch over to the pencil instead of turning drawing off.
-    if (fabricCanvas.value.isDrawingMode && !isEraserModeOn.value) {
-      fabricCanvas.value.isDrawingMode = false
-      return
-    }
-    isEraserModeOn.value = false
-    if (fabricCanvas.value) {
-      fabricCanvas.value.isDrawingMode = true
-    }
-    const brush = new PSBrush(fabricCanvas.value)
-    brush.pressureManager.fallback = 0.5
-    // PSBrush drops BaseBrush's round cap/join defaults (its initialize
-    // doesn't call super), so restore them or strokes get flat ends.
-    brush.strokeLineCap = 'round'
-    brush.strokeLineJoin = 'round'
-    lockBrushToFirstPointer(brush)
-    fabricCanvas.value.freeDrawingBrush = brush
-    _resetColor()
-    _resetPencil()
-  }
-
-  // Eraser tool — mirror of onAnnotateClicked. Installs the v2 EraserBrush
-  // (destination-out). Type-selectivity is handled per-object by the
-  // `erasable` flag (IText is non-erasable), not by z-order.
-  const onEraseClicked = () => {
-    showCanvas()
-    if (isEraserModeOn.value) {
-      isEraserModeOn.value = false
-      if (fabricCanvas.value) fabricCanvas.value.isDrawingMode = false
-      return
-    }
-    isEraserModeOn.value = true
-    isShapeMode.value = false
-    if (fabricCanvas.value) {
-      fabricCanvas.value.isDrawingMode = true
-      const brush = new EraserBrush(fabricCanvas.value)
-      brush.strokeLineCap = 'round'
-      brush.strokeLineJoin = 'round'
-      lockBrushToFirstPointer(brush)
-      fabricCanvas.value.freeDrawingBrush = brush
-      _resetEraserWidth()
-    }
-  }
-
-  const _resetEraserWidth = () => {
-    if (!fabricCanvas.value?.freeDrawingBrush) return
-    // Reuse the pencil-width preference (same mapping as _resetPencil).
-    const converter = { huge: 30, big: 20, medium: 10, small: 4, tiny: 2 }
-    fabricCanvas.value.freeDrawingBrush.width = converter[pencilWidth.value]
-  }
-
   const onObjectAdded = obj => {
     if (silentAnnotation) return
     let o = obj.target ? obj.target : obj.targets[0]
@@ -1019,52 +912,6 @@ export const useAnnotation = ({
 
   const setAnnotationCanvasDimensions = (width, height) => {
     fabricCanvas.value.setDimensions({ width, height })
-  }
-
-  const setAnnotationDrawingMode = isDrawingMode => {
-    if (isDrawingMode) {
-      isShapeMode.value = false
-      // Coming back to the pencil from the eraser: the eraser swapped
-      // freeDrawingBrush for an EraserBrush, so restore the pressure brush
-      // (consumers that toggle drawing through this entry point — e.g.
-      // PreviewPlayer — never re-create a PSBrush themselves).
-      if (fabricCanvas.value?.freeDrawingBrush instanceof EraserBrush) {
-        isEraserModeOn.value = false
-        const brush = new PSBrush(fabricCanvas.value)
-        brush.pressureManager.fallback = 0.5
-        brush.strokeLineCap = 'round'
-        brush.strokeLineJoin = 'round'
-        lockBrushToFirstPointer(brush)
-        fabricCanvas.value.freeDrawingBrush = brush
-        _resetColor()
-        _resetPencil()
-      }
-    } else if (isEraserModeOn.value) {
-      // The eraser runs in drawing mode. A stray "leave drawing" — typically
-      // the isDrawing watcher firing async as the user switches pencil→eraser
-      // (onEraseClicked set isDrawing=false) — must not tear it down, or the
-      // eraser stops working and its ring cursor reverts to the default.
-      return
-    }
-    fabricCanvas.value.isDrawingMode = isDrawingMode
-  }
-
-  const toggleShapeMode = () => {
-    if (isShapeMode.value) {
-      isShapeMode.value = false
-      return
-    }
-    isShapeMode.value = true
-    // Mutex with the freehand drawing mode owned by the composable.
-    // Consumers (PreviewPlayer / PlaylistPlayer) own `isDrawing` /
-    // `isTyping` refs and clear them via a watcher on `isShapeMode`.
-    if (fabricCanvas.value) {
-      fabricCanvas.value.isDrawingMode = false
-    }
-  }
-
-  const setShapeTool = shape => {
-    currentShape.value = shape
   }
 
   // Suppress object selection while shape mode is active so a mousedown
@@ -1546,6 +1393,32 @@ export const useAnnotation = ({
       canvasWrapper.value.style.display = 'block'
     }
   }
+
+  // Pencil / eraser / shape / text tool transitions and brush configuration
+  // live in their own module; the facade re-exposes them unchanged.
+  const {
+    onChangePencilColor,
+    onChangePencilWidth,
+    onChangeTextColor,
+    _resetColor,
+    _resetPencil,
+    resetPencilConfiguration,
+    onAnnotateClicked,
+    onEraseClicked,
+    setAnnotationDrawingMode,
+    toggleShapeMode,
+    setShapeTool
+  } = useDrawingTools({
+    fabricCanvas,
+    isEraserModeOn,
+    isShapeMode,
+    currentShape,
+    pencilColor,
+    pencilWidth,
+    textColor,
+    showCanvas,
+    localPreferences
+  })
 
   // Helper to get the current preview (needed by updateAnnotationsInStore)
   let currentPreview = () => null
